Fix bug with starting level
There was a bug whereby on mobile (tested on Android phone) levels would fail to start properly (bomb would not start rotating or animating its fuse flare). It seems this issue was related to failing to find a reference to the bomb Game Object using FindGameObjectsWithTag. Using the reference returned from Instatiate is much more reliable.